pyproject.toml
[project]
name = "tabletop-planning-env"
version = "0.1.0"
description = "OpenEnv tabletop replanning benchmark for embodied task planning"
requires-python = ">=3.10"
dependencies = [
    "mujoco>=3.0.0",
    "robosuite>=1.4.1",
    "numpy>=1.24.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.0.0",
    "gymnasium>=0.29.0",
    "openenv-core>=0.2.1",
]

[project.optional-dependencies]
train = [
    "trl>=0.8.0",
    "transformers>=4.40.0",
    "torch>=2.2.0",
    "datasets>=2.18.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"
